for testing: 'deleting a blog with existing id responds 204'

+++ part4/controllers/blogs.js
@@ -1,7 +1,7 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-const User = require('../models/user')
-const jwt = require('jsonwebtoken')
+
+const userExtractor = require('../utils/middleware').userExtractor
 
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog
@@ -18,23 +18,10 @@ blogsRouter.get('/:id', async (request, response) => {
     }
 })
 
-const getTokenFrom = request => {
-    const authorization = request.get('authorization')
-    if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
-        return authorization.substring(7)
-    }
-    return null
-}
+blogsRouter.post('/', userExtractor, async (request, response) => {
+    const { title, author, url, likes } = request.body
 
-blogsRouter.post('/', async (request, response) => {
-    const { title, author, url, likes, id } = request.body
-
-    const token = getTokenFrom(request)
-    const decodedToken = jwt.verify(token, process.env.SECRET)
-    if (!decodedToken.id) {
-        return response.status(401).json({ error: 'token missing or invalid' })
-    }
-    const user = await User.findById(decodedToken.id)
+    const user = request.user
 
     if (!title && !url) {
         response.status(400).end()
@@ -55,9 +42,22 @@ blogsRouter.post('/', async (request, response) => {
     response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {//with express-async-errors
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {//with express-async-errors
+
+    const user = request.user
+    const blog = await Blog.findById(request.params.id)
+
+    if (blog) {
+        if (blog.user.toString() === user._id.toString()) {
+            await Blog.findByIdAndRemove(request.params.id)
+            response.status(204).end()
+        } else {
+            return response.status(401).json({ error: 'you don\'t have permission to perform this action' })
+        }
+    } else {
+        response.status(204).end()
+    }
+
 })
 
 blogsRouter.put('/:id', async (request, response) => {
